flask_app/requirements.txt
requests==2.25.1
beautifulsoup4==4.9.3
openai==0.27.2
flask==2.2.3
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-api-core==2.11.0
Jinja2==3.1.2
gunicorn==20.1.0
Pygments==2.14.0
Flask-Cors==3.0.10